+++ src/main/kotlin/us/frollo/frollosdk/extensions/APIExtension.kt
@@ -36,6 +36,7 @@ import us.frollo.frollosdk.model.api.goals.GoalResponse
 import us.frollo.frollosdk.model.api.images.ImageResponse
 import us.frollo.frollosdk.model.api.messages.MessageResponse
 import us.frollo.frollosdk.model.api.reports.AccountBalanceReportResponse
+import us.frollo.frollosdk.model.api.reports.CashflowReportResponse
 import us.frollo.frollosdk.model.api.reports.ReportsResponse
 import us.frollo.frollosdk.model.api.shared.PaginatedResponse
 import us.frollo.frollosdk.model.api.statements.Statement
@@ -247,6 +248,16 @@ internal fun ReportsAPI.fetchTagReports(
     return transactionTag?.let { tag ->
         fetchReportsByTag(tag, queryMap)
     } ?: fetchReportsByTag(queryMap)
+}
+
+internal fun ReportsAPI.fetchCashflowReports(
+    period: TransactionReportPeriod,
+    fromDate: String,
+    toDate: String
+): Call<CashflowReportResponse> {
+
+    val queryMap = mutableMapOf("period" to period.toString(), "from_date" to fromDate, "to_date" to toDate)
+    return fetchCashflowReports(queryMap)
 }
 
 // Bills