+++ src/sharedTest/kotlin/us/frollo/frollosdk/model/TestCashflowReportData.kt
@@ -1,0 +1,3 @@
+package us.frollo.frollosdk.model
+
+class TestCashflowReportData